[#91] Refactor: application 및 infrastructure 하위 패키지 정리
